Default to the ORU outcome for deaths registration abroad
Now that all the countries that were supposed to be transitioned to ORU outcomes
have been transitioned to ORU outcomes, there are only four distinct groups of
countries that influence the flow:

1) Commonweatlh countries - lead to a separate outcome early on
2) No embassy countries - lead to the no embassy outcome
3) North Korea - exceptional, handled by ORU unless the user is there and birth/death
   occured in N.Korea, in which case the local embassy handles it.
4) Countries handled by Overseas Registration Unit

Those four groups cover all the countries visible in the countries drop down list.

Previously, if a new country was added to the dropdown list, it was leading to the
embassy outcome, which used to be the default way to handle registrations. It now
makes more sense to default to the ORU outcome. Therefore we treat each country as
an ORU country unless it's in one of the lists (1-3).

This change should not have any impact to the current tool's behaviour.

The list will be removed in a later commit.