Remove ORU_TRANSITION_EXCEPTIONS and references to it
N. Korea is the only country in the ORU_TRANSITION_EXCEPTIONS. I think the
constant name is confusing, as the ORU transition is over and we should not be
bothered by the fact that there was a gradual transition. N. Korea is an exceptional
country and it is very unlikely another country will follow a similar procedure.

Therefore I decided to explicitly check for N.Korea in the code logic to make it
easier to understand it.