Indent check-uk-visa outcomes
NOTE. I've not indented the '_b1_b2_visa_exception.govspeak.erb' partial as
doing so results in some of the rendered outcomes changing. This is because
this partial is rendered at the end of a number of sentences and indenting it
results in two new spaces appearing in those sentences. While these spaces
shouldn't effect the HTML that's rendered I still felt that this change was out
of scope of what I'm doing at the moment.