Rename register-a-birth/embassy_result to north_korea_result
Only when the birth has occured in North Korea and the user is there
the logic leads to this outcome. The outcome also has some Korea-specific
phrases, so make it explicitly clear that this outcome is not generic.

It is also very unlikely any other country will switch back from ORU
to embassy process, so this should be alright.