Update links manually to student finance forms
We've been asked to update the links to student forms as the stakeholder
is changing the CMS that is serving the forms.

Some of the old links provided did not match exactly, but I was able to
deduct the needed new link based on the last part of the URL and checking
the actual pdfs in the browser.