Move formatting of minimum_hourly_rate to templates
Formatting is a view concern rather than a flow concern.

I've moved the formatting of `minimum_hourly_rate` from the flow to the ERB
templates for am-i-getting-minimum-wage and minimum-wage-calculator-employers.

The `MinimumWageCalculator#format_money` and `OutcomeHelper#format_money`
methods both remove the zeros for whole numbers. The
`OutcomeHelper#format_money` also adds separators to large numbers and prepends
the currency symbol.

I've updated a number of the test artefacts that have changed now that larger
numbers include a separator.