Consistently use `next_node` blocks
Rather than a mix of `next_node` and the `option 'key' => 'destination'`
shortcuts.

I've gone with `next_node {}` as it's the more flexible of the two approaches.
We can always change from the shortcut method to the block method, but we can't
always go the other way as we sometimes rely on calculating/settings things in
the `next_node` blocks.