Instantiate the calculator in the first question of the flow
I'm planning to set the responses to questions on the calculator when we
receive those responses, rather than in a single question later in the flow
(:how_much_are/were_you_paid_during_pay_period?).

This should avoid us having to store the responses on the state object.

I've had to override the `date=` method so that I can call
`minimum_wage_data_for_date(@date)` and set the `@minimum_wage_data` variable.
This was previously happening in the constructor but I can no longer rely on
that as we're constructed the object before having the `date` available. I'm
sure this can be improved but I think it's probably good enough for now.