Use ActionDispatch::TestRequest instead of Hash w/ stubbed method
This feels more realistic and has the advantage of not stubbing a non-existent
method (i.e. `Hash#query_parameters`).

Note that `ActionDispatch::TestRequest#query_parameters` returns `{}` by
default so we no longer need to stub it in those cases.

There's probably a way to set `query_parameters` on the request rather than
stubbing the method, but I couldn't immediately see a way.