Avoid stubbing non-existent methods in holiday entitlement calc
The `HolidayEntitlement` calculator uses `method_missing` (argh!) to implement
magic methods to format attributes on the calculator. This commit implements
the corresponding `respond_to?` method which you are supposed to implement if
you use `method_missing`.

I want to configure Mocha to prevent stubbing of non-existent methods, but the
test for the formatting mechanism described above were making use of a
non-existent method `#foo`. This commit defines the method so that it can
safely be stubbed.

Finally I've updated the regression test checksums for the relevant flow.