Introduce MinimumWageCalculator#apprentice_elligible_for_minimum_wage?
Policy related logic should be encapsulated in the `MinimumWageCalculator`
rather than the flow.

NOTE. I've reversed the condition in the `next_node` block of the
`:were_you_an_apprentice?` question. This was to avoid either using `if !` (or
`unless`) or having to negate the new method (e.g.
`#apprentice_not_elligible_for_minimum_wage?`).