Introduce new DateRange class
This is a value object [1] representing a range of dates. It can be
instantiated using a start date and either an end date or a duration. If
neither an end date or a duration is specified then it is assumed that the
range never ends.

The start and end dates are converted to a date (in case they aren't already).
The duration should be an object that works with date addition, e.g. a
`Numeric` number of days or an instance of `ActiveSupport::Duration`. This will
ensure that ranges including 29th February in a leap year are correctly handled.

Since it's a value object, I've implemented equality-related methods so that
`DateRange`'s with the same start and end dates are equal/equivalent to each
other. These implementations are based on those in the `values` gem [2].

I'm planning to use this in the new calculate-your-part-year-profit flow, but I
expect it could be used in some of the existing flows.

[1]: http://c2.com/cgi/wiki?ValueObject
[2]: https://rubygems.org/gems/values