Introduce a TaxYear class
This is a value object [1] representing the range of dates of a UK tax year.
It is instantiated using the calendar year in which the tax year begins. Factory
methods for the current tax year and the tax year for a specific date are also
provided.

I'm planning to use this in the new calculate-your-part-year-profit flow, but I
expect it could be used in some of the existing flows.

[1]: http://c2.com/cgi/wiki?ValueObject