Introduce PartYearProfitCalculator class
This is intended to capture the policy logic for the upcoming
calculate-your-part-year-profit smart answer.

The intention is that the flow will use the attribute setter methods, but as a
convenience for the tests, I've included `ActiveModel::Model` which implements
a constructor accepting a hash of attributes and values.

The logic in this commit represents only the "happy path". I plan to extend
this logic in subsequent commits.