Fail regression tests if new outcomes files are generated
Currently the tests fail if outcome files have been removed or changed. This
is because `git diff --stat` is used in the implementation. `git diff` does not
list untracked files.

This commit adds a check for newly generated outcome files using
`git status --short -- <path>` and explains what needs to be done to resolve
this test failure.