Revert "An attempt at fixing intermittent test failure"
This reverts commit 1d78712ad821f58909c38f3e1d2e8a7ed9205f45.

It turns out that this didn't fix the problem it was intended to fix i.e. I've
seen the `Capybara::Poltergeist::JavascriptError` exceptions mentioned in #1624
occur again on my local machine when running the tests.

Also this commit seems to have had the undesirable side-effect of stopping the
WebMock stubbing from working when running flow integration tests in isolation
- at least it does on my machine. I don't really understand why the stubbing
seems to work when running all the tests or why running an integration test in
isolation worked on @chrisroos' machine.

I spent a bit of time investigating the above, but didn't really get anywhere.
Given that the fix didn't solve the original problem anyway, it doesn't seem
very worthwhile putting more time into it.