Preload entire app before forking Unicorn workers.
smart-answers is CPU hungry, especially during app
startup. It's preloading the answers and some of
them take a long time to load. We have seen an
issue where application was slow to start.

smart-answers does not have a database, so we don't
need to deal with the before/after fork dance.

Setting this to true reduces the start up time for
starting up the Unicorn worker processes. This
uses CoW to preload the application before forking
other worker processes.