Optionally mark country_name_lowercase_prefix as html_safe
This ensures we continue to generate the same govspeak when the
`ceremony_country` is "cote-d-ivoire". If we don't mark the resulting string
("Cote d'Ivoire") as HTML safe then the outcomes contain "Cote d&#39;Ivoire"
instead of "Cote d'Ivoire". The "&#39;" is then further escaped which means
that we end up seeing that on the page instead of the single quote.