Convert entitled_to_sick_pay outcome
This conversion has resulted in a couple of whitespace changes to the test
artefacts, although I'm going to commit those separately to reduce the noise in
this commit:

1. A leading space has been removed from before single digit numbers and Total
SSP in the table of SSP figures.

I checked the following URL to ensure that this whitespace change had no effect
on the rendered outcome:

/calculate-statutory-sick-pay/y/none/no/no/2013-04-02/2013-04-10/no/before_payday/weekly/2000.0/30/1,2,3,4,5

2. A newline been removed from above "SSP can be paid for up to ...". I can't
work out how to add this back without affecting other artefacts.

I checked the following URL to ensure that this whitespace change had no effect
on the rendered outcome:

calculate-statutory-sick-pay/none/no/no/2013-04-02/2013-04-10/no/eight_weeks_more/irregularly/2013-03-31/2013-01-31/4000.0/1,2,3