Avoid calculate block in :current/past_accommodation_charge?
The `accommodation_charge` variable is used in the call to
`#accommodation_adjustment` in the `:current/past_accommodation_usage?`
questions. The `#accommodation_adjustment` method already converts the `charge`
argument to a `Float` which means we can avoid doing that in this `calculate`
block, and simply rely on using `save_input_as` instead.