Introduce MinimumWageCalculator#any_overtime_hours_worked?
Policy related logic should be encapsulated in the `MinimumWageCalculator`
rather than the flow.

NOTE. I've reversed the condition in the `next_node` block of the
`:how_many_hours_overtime_do/did_you_work?` questions. This was to avoid either
using `if !` (or `unless`) or having to negate the new method (e.g.
`#no_overtime_hours_worked?`).