UK-issued CNI is now accepted by Greek town halls
The Greek Ministry has now agreed for Greek Town Halls to accept the Domestic
CNI issued by the UK Superintendent Registrar for marriage in Greece without
it having to be exchanged by FCO.

There is a very high volume of British nationals who marry in Greece over the
summer. This is a significant change and users should be notified that the
extra financial cost of exchanging the CNI is no longer needed.