Add Smartdown support for the `generate-checksums-for-smart-answer` command
This automatically includes all .txt files under `lib/smartdown_flows/<name>`
except for scenarios.

Worth noting is that the script now silently ignores paths that do not exist.
This is because it does not know if it was given a Smartdown or a ruby format.
To compensate for this it now informs the user if no files at all were detected.