Move over student-finance-forms start page and the first question
Note that "type_of_student is 'eu-full-time'" is not leading to the
what_year_eu_fulltime node. It appears that what_year* nodes ask the same
question, moreover, the what_year.txt smartdown node has all the logic to handle
all input variants.

Looking at eded6b648e895b0768c5b20719c5c7cf7c412ef0 and
11280c872e51aa81d7104ea581fa5a2fb9ff82eb it seems that there has been some
effort to simplify logic in what_year.txt by splitting it out to several nodes.
Required logic was moved to the new what_year_* nodes, but it was not deleted
in what_year.txt nodes.

For the purposes of the conversion it is going to be simpler to handle the flow
using the logic in a single "what_year" node.

This commit is a part of the student-finance-forms conversion from smartdown to
ruby smartanswers.