Regenerate student-finance-forms-questions-and-responses.yml
This needs to be regenerated, because the smartanswer has been migrated from
smartdown to the ruby flow format. Question nodes have changed to have a
trailing question mark and the order of questions in the yml file has changed.

Due to the latter it is a bid more difficult to read the diff, but the data
should not have changed.