Avoid using `contact_method_key` in outcome_ss_marriage
Relying on the `contact_method_key` meant that I'd have to retain the
duplication between the various 'appointment_links' Phrases in
marriage-abroad.yml and the '_contact_method.govspeak.erb' partial.

Checking that `contact_method_key == :embassies_data` is equivalent to the
`case` statement in '_contact_method.govspeak.erb' falling through to the
`else` clause. So, instead of relying on a value of `contact_method_key` I can
pass the text we want displayed to the '_contact_method.govspeak.erb' partial
and have that decide whether to display it.

This change doesn't break any of the regression tests for the
outcome_ss_marriage outcome.