Fix temporary directory clean-up in SmartAnswerFilesTest
`FileUtils.rm_f` doesn't delete directories.

Even though a directory was being left behind in `lib/smart_answer_flows`, this
wasn't causing any problems. When we started generating a directory tree for
smartdown flows in [1] this caused an error (see below) in
`SmartdownScenariosTest`:

    .bundle/gems/gems/smartdown-0.15.1/lib/smartdown/parser/directory_input.rb:77:in `read': No such file or directory @ rb_sysopen - lib/smartdown_flows/smartdown-answer/smartdown-answer.txt (Errno::ENOENT)

Presumably the problem was not occurring on CI, because it uses a fresh
directory for every build.

[1]: 470ff8421400fb80c462632b2106ae1b0901507c