DRY up smart_answers_controller_test
I'm in the process of removing support for Outcome text to be stored in YAML
locale files.

The previous commit updated the 'sample' flow in the
smart_answers_controller_test to use ERB templates for all outcomes. That
change resulted in two almost identical tests for the text representation of
Smart Answer outcomes. I'm removed the newer of those tests, along with the
associated 'outcome_with_erb_template' outcome.