Remove unnecessary call to use_template?
We're now assuming that all Outcomes are stored in ERB templates so this check
is redundant.

I've also removed the "Outcome has no title if using the OutcomePresenter" test
that has started failing since this commit.

The test was introduced in f55517cbc15e7ce7b6ba1c00e2b28a86df65e61c and was
added to demonstrate that titles were optional for Outcomes, even though they
were always set for other types of Node. The OutcomePresenter retains this
behaviour for ERB templates, and that behaviour is tested in the
OutcomePresenterTest.