Fix bug in pre-fill when changing answer to postcode question
The value in the text field for a postcode question was not pre-filled when
clicking the change answer link on a later question.

This bug was present for smartdown flows prior to the introduction of
`postcode_question` for smart answer flows, but this question type was only
used in a single draft smartdown flow, `landlord-immigration-check`.

However, it was easier to write a test to demonstrate the problem in a
smart answer flow once I'd introduced `postcode_question` to smart answer
flows.

The issue was that `CurrentQuestionHelper#previous_response_key` already
correctly defaulted to `:previous_response` if no number was given, but
the call to `CurrentQuestionHelper#prefill_value_for` in the template was
second-guessing this by defaulting `number` to `1`. Unfortunately this
meant that `#prefill_value_for` then looked for a `:previous_response_1`
parameter which did not exist.

I'm not 100% sure this fix would work with multiple `postcode_question`s
per page, but since we don't currently support multiple questions per page
for smart answer flows, this is moot.