Introduce LandlordImmigrationCheckCalculator
This is pretty much a direct copy of
`SmartdownPlugins::LandlordImmigrationCheck` and its associated unit test.

The only difference is that we're expecting the smart answer flow to pass in
the postcode as a `String` and not an object with a `value` attribute.

This calculator will take over from the `SmartdownPlugins` version when we
convert the smartdown flow `landlord-immigration-check` to a smart answer flow.

The code in the `valid_postcode` method is a bit horrible, but it seems better
to keep the code as similar as possible in this commit. I hope to improve it
later.