next_nodes for regression test should be questions not pages
Smartdown flows have question pages as top-level nodes each of which has one
or more questions.

We were already correctly handling this for the `current_node` [1], but the
`next_node` was incorrectly being set to a question *page* name not a question
name.

I've updated the relevant files for the `landlord-immigration-check` which is
currently the only smartdown flow with regression tests.

[1]: https://github.com/alphagov/smart-answers/blob/5da729a56e7a3e673e8e789cd90a8c424fcfadc8/script/generate-responses-and-expected-results-for-smartdown-flow.rb#L26