Convert landlord-immigration-check from smartdown to smart answer
I ran the following command:

    rails r script/convert-smartdown-flow-to-smart-answer-flow.rb landlord-immigration-check

Note that this "automatic" conversion has left a couple of TODO items which
I intend to fix in a subsequent commit.

Note also that the question symbols do not end in a question mark. This is so
that the regression test assertion that all nodes are exercised will pass
without needing to make changes to the input files. I plan to add the question
marks in a later commit.