Fix TODO items in converted landlord-immigration-check flow
At this point the regression tests all pass with *no changes* to the input
files. This should give us a lot of confidence that nothing has been broken in
the conversion process.

The regression tests run against the smart answer version of the flow in
preference to the smartdown version, even though both exist.

I have updated the regression test checksums, but note that since both versions
of the flow exist files for both are included in the checksum files.