next_nodes for regression test should be symbols not strings
I want the `questions-and-responses` and `responses-and-expected-results` files
which are used as inputs to the regression tests to remain the same when we
convert a smartdown flow to a smart answer flow.

Previously we were storing node names as symbols for smart answer flows, but as
strings for smartdown flows. This commit brings the latter into line with the
former.

I've updated the relevant files for the `landlord-immigration-check` which is
currently the only flow with regression test input files in this form. It looks
as if the equivalent files for the `student-finance-forms` smartdown flow were
updated in [1] after the flow had been converted.

[1]: 2f41fa9a68803460805b320b172ee3442ae98ef8