Run govuk-lint-ruby on Jenkins
Using the `--diff --cached` options means that this should only lint files that
have been touched in this branch.

I've chosen the `clang` format as that matches the Jenkins configuration in
signonotron2[1], slimmer[2] and whitehall[3]. It's also a format that makes it
easy to see the errors that have been introduced.

I've not added the `html` format, which is present in the whitehall and
signonotron2 config as I'm not convinced we need it.

I'm not constraining the directories that Rubocop searches in for violations
as, at this early stage, I'd rather restrict them on a case by case basis as
necessary.

[1]: https://github.com/alphagov/signonotron2/blob/master/jenkins.sh
[2]: https://github.com/alphagov/slimmer/blob/master/jenkins.sh
[3]: https://github.com/alphagov/whitehall/blob/master/jenkins.sh