Transform outcome_netherlands to outcome_marriage_via_local_authorities
This is in preparation of transitioning several more countries to this outcome.

Brings the outcome in line with other outcomes by interpolating the
country name with prefix in the title (same as CNI and affirmation outcomes) and
clarifying what local authorities are. If the user is in the UK, advise to
contact the embassy instead of local authorities in the country.

Remove partner naturalisation paragraph when partner is British

The outcome was incorrectly displaying information about
partner naturalisation when the partner is British. This piece of information
is irrelevant when both partners are British.

Users should get a travel advice if they're not in the country.