Unify a sentence in overseas passports
A version with 'first name and last name' of this sentence is repeated a few
times in this Smart Answer. After consulting with a content designer I am making
this sentence consistent with the others. I am planning to reduce duplication
of the "Making your application" section and this should make it easier.

None of the regression test artefacts seems to cover the case where this
sentence is used, therefore non of the artefacts have changed.