Reduce content duplication in the "Making your application" section.
Extract commonly used paragraphs into partials and reuse them.

This introduces insignificant whitespace changes to outcomes and changes one
sentence from "You’ll get an email..." to "You will receive an email..." so that
the same partial can be used. It is OK with a content designer.

Rationale: even though content is not code, duplication of it is not great.
When change requests come in, they usually have before/after text. If the text
is duplicated, it makes it hard to trace back where the change needs to be made.

It also reduces the text to code ratio in the making_your_application template,
making it easier to follow logic flows.