Extract OS marriage in Italy to its own outcome
Opposite-sex marriages in Italy have plenty of exceptions. That means a lot of
extra ifs in the generic CNI outcome. Splitting it out duplicates some of the
CNI logic, but also removes those extra ifs in the already big CNI outcome.

Whitespace changes in generated Italy outcome artefacts are not significant.