Modify 'What happens next' for marriage in Poland
When a user is residing in Poland, the 'What happens next' section is changing
to remove all references to Notary Public.

For users in Poland include the section about how early do users need to
provide the CNI that got recently added to the outcome for UK users. Make sure
the "What happens next" title is not repeated in the outcome for users in
Poland.