Improve residency requirements explanation for marriage in UAE
Only show the possibility to get married _in the embassy_ if a user is resident
in UAE (both partners need to be residing there for this option).

If a user is not resident in UAE inform them that at least one partner needs to
be resident in UAE to get married in the country.