Improve residency requirements explanation for marriage in UAE
Do not show the possibility to get married in the embassy unless a user is
resident in UAE (both need to be residing there for this option).

If a user is not residing in UAE, make sure to show the minimum requirement for
marriage (at least one partner). We don't need to show the latter in the case
the user is resident in UAE because this condition is already satisfied.