Add support for businesses that have recently started trading
This change means that we no longer send anyone to the unsupported outcome. The
two paths that have changed are:

1. If you don't have 12 months of accounts then the assumption is it's because
you started trading within that 12 month period. We now ask when you started
trading so that we can calculate your pro rata profit from the period you
started trading.

2. If you've ceased trading and weren't trading before the beginning of the
relevant accounting year then we now ask when you started trading so that we
can calculate your pro rata profit from the period you started trading.

NOTE. I'd hoped to have been able to set the `from` and `to` in the
`:when_did_you_start_trading?` question to dates relative to the tax year in
which Tax Credits award ends. Unfortunately, these blocks aren't evaluated in
the context of the state so that's not immediately possible.