Remove Question::Date#default
This was added in 1ac12b6b0be70aa5a816e707e240cbfdf67d5692 to allow a default
date to be specified for the maternity Smart Answer. It no longer appears to be
in use so I think it's safe to remove.

I've had to amend the _question_date partial so that it no longer calls
`question.default`. This shouldn't cause any problems as `question.default`
would have always been returning `nil`.