Remove Question::Date#defaulted_year?
The `#default_year` method will return either `nil` or the result of the block
stored in `@default_year_func` if set. Using this means that we can remove the
`#defaulted_year?` method.

This was added in 8b58286e13005724e4cb4a91ab0a16bcdc9ae8b4 but it's not clear
why the additional method might be necessary.