Add tests for Question::Date#parse_input
I'm looking to make some small changes to this method and wanted some tests in
place first to give me some confidence.

There appears to be some overlap between these tests and some other tests in
this file that use `#transition` to indirectly test the `#parse_input` method.
I'm happy to live with that duplication for now although I may remove it later.