Avoid rescuing all StandardError exceptions in parse_input
I'm fairly confident that the intention is to raise an `InvalidResponse`
exception when the date can't be parsed. We raise that exception when we detect
an invalid date, or when `Date.parse` or `Date.new` raises an `ArgumentError`.
Now that I have tests in place around this method I'm able to make the `rescue`
statement stricter which should hopefully make the code fail fast if it's truly
used in exceptional circumstances.