Simplify parsing of dates supplied as hashes
This is probably somewhat subjective but I find this implementation slightly
easier to understand.

NOTE. I've removed the "Please enter a complete date" error message as it wasn't
serving any purpose. This string would've become the `#message` of the
`InvalidResponse` instance which is used as a key to lookup the actual error
message in the YAML file associated with the Smart Answer. The "Please enter a
complete date" string isn't a key and so `NodePresenter#error` would've fallen
back to finding either the default error message for the question or the default
error message for all flows.