Give nodes a reference to their parent flow
I'm going to need the functionality in `Outcome#template_directory` in
`StartNodePresenter` and we will eventually want it in question node
sub-classes. This method needs to know the name of the flow in order to
calculate the path to its templates. Building nodes with a reference to their
parent flow seems the nicest way of achieving this.