Move dates from flow to calculator
I hope this makes it clearer that the dates we use in the `from` and `to`
blocks of the 'when_did_you_stop_trading?' and 'when_did_you_start_trading?'
questions are related to the dates used in the
'when_did_your_tax_credits_award_end?' question.

See commit with subject "Set from and to date range for
when_did_your_tax_credits_award_end?" for the significance of the 2015-01-01
and 2018-12-31 dates.

See commits with subjects "Add support for businesses that have stopped
trading" and "Add support for businesses that have recently started trading"
for the reasons behind the other dates.