Split out bus pass calculation as a separate flow branch
Currently state pension age outcomes from from this calculator show a paragraph
about bus pass eligibility age. However it is not very easy to find as it is
at the bottom and only one of many paragraphs. A lot of users find it confusing.

A lot of local council and other authority websites link to this calculator
for bus pass age calculation (that's why we rejected an idea of having a
completely separate calculator), therefore making this option explicit should
make it easier for users. The new outcome has only information about the bus
pass age and a link back to State Pension age calculation (ideally we'd link to
an outcome, but we don't know user's gender).

I opted in for some duplication to introduce a bus_pass_dob? question. I could
have reused age_dob? question, but inelegant hacks are needed to workaround the
fact that an instance of StatePensionAmountCalculator is initialized there
and it requires gender to be passed in.