Refactor old student maintenance grants and loans calculations
This gets rid of the comments and brings the implementation inline with the
new calculations style. I decided to get rid of comments because now behaviour
is covered by unit tests that explain logic in descriptions. One comment also
had an incorrect value that was out of sync with code (£9.90 vs £9.59).