Declare many methods in StatutorySickPayCalculator as private
StatutorySickPayCalculator public API is rather extensive, making significant
changes to it a bit harder. Declare all methods that are not used outside of
this class as private. I didn't feel comfortable enough to re-factor the tests
yet and where tests are testing methods that are now private I used
Object#send. This is an anti-pattern, but I see it as exposing a symptom of
poor class API design that was there already rather than a newly introduced
problem.

@floehopper: I'm not completely convinced about the changes in this commit, but
I've decided to apply it anyway, because I don't think it does any harm and it
should make merging of some other changes less painful.