Declare many methods in StatutorySickPayCalculator as private
StatutorySickPayCalculator public API is rather extensive, making significant
changes to it a bit harder. Declare all methods that are not used outside of
this class as private. I didn't feel comfortable enough to re-factor the tests
yet and where tests are testing methods that are now private I used
Object#send. This is an anti-pattern, but I see it as exposing a symptom of
poor class API design that was there already rather than a newly introduced
problem.