Declare some methods in StatePensionTopup Calculator as private
Declare all methods in StatePensionTopup that are not used outside of this
class as private.

There are tests testing methods that are now private so I used Object#send to
keep them working. This is an anti-pattern, but I see it as exposing a symptom
of poor class API design that was there already rather than a newly introduced
problem.