Require parent contact details form in three countries
Bangladesh, India and Pakistan now require parents applying for their child's
first passport to fill in the parent contact details form.

Note that the numbering in the bullet points is off by one in some cases in
Pakistan. This could by avoided by introducing more conditional branches, but
I opted in to simply ignore it, because when markdown renders ordered lists
it simply strips provided numbers and uses HTML's <ol> element that shows the
correct numbers to users.