Differentiate between Israeli travel documents
We've had letters of complaint from customers that the Visa Checker tool is not
clear for holders of Israeli travel documents in lieu of passports. Israeli
passport holders are non-visa nationals, so in general only require entry
clearance if they are coming here for longer than 6 months. Israeli-issued
travel documents are issued to non-Israeli citizens (like Palestinians) and
the holders are visa nationals so require a visa to come to the UK for any
reason.

Add an Israel-specific question that collects information about the user's
travel document type and show the corresponding outcome.