Allow women aged 62 to make state pension top ups
Add a missing rate for 62 year old women. Previously the threshold for female
users in the calculator was 62 years. According to the department it is
incorrect and women born between 13 October 1952 and 5 April 1953 (inclusive)
should be able to top up when they are 62.

The tests have been changed to illustrate realistic scenarios provided by the
department, but they are not passing at the moment because there is another
change required that I'm going to commit next.