Fix indexable content for outcome nodes in flow registration
This was broken when we switched over to use only ERB templates for outcomes.

The unit test continued to pass, because it used `NodePresenter` for *all* node
types and the outcome node content was still present in the fixture YAML file.

This commit changes the code to use `OutcomePresenter` for outcome nodes and
moves the outcome node content into ERB templates.

Also note that only question nodes have hints, so I've updated the content and
assertions accordingly.