Send `content_id` to panopticon
By adding `FlowRegistrationPresenter#content_id` we are sending the
content_id of a flow to panopticon when running `rake
panopticon:register`.

The way this works is that `content_id` is in the list of optional
values to forward to panopticon in the "panopticon registerer". This
means that if you call `GdsApi::Panopticon::Registerer#register` with
an object that responds to `content_id` it will send the value to
panopticon automatically. :magic: