Refactor RegisterableSmartAnswers and improve test
Currently smartdown and erb-smart answers are de-duplicated in
`RegisterableSmartAnswers`. We assume this "happened because when there
were plans to convert Smart Answers from Ruby to Smartdown, which may
have resulted in duplicates during the transition." (@chrisroos)

We can clean up the code to only merge and decorate the flows. Now the
purpose of the class is clearer we can add some more robust tests.