Update passport application advice for users in Kyrgyzstan
Kyrgyzstan applications are now fulfilled by the Embassy, not the application
centre.

I've struggled to work out how to make this change correctly.

Originally @tadast attempted to make the desired changes in these two commits:

* 8286a98d0456c363abb88d6233584fdda0a05721
* 831677a1f5c7364352b16337f930803269dc06c7

However, in fact-checking FCO noticed that the whole "Making your application"
section had disappeared. Initially I thought that this was because @tadast had
forgotten to disable the `application_office` flag in `passport_data.yml`. But
when I tried doing this, although the "Making your application" section did
reappear, it was still missing a bunch of sentences/paragraphs.

I discussed the problem with Abigail Waraker and she thinks that "Burundi" works
the way we now want "Kyrgyzstan" to work, so I've made the changes in this
commit to bring "Kyrgyzstan" into line with the "configuration" that "Burundi"
uses.

I'm not very confident that this is the right/best way to make this change, but
since it needs to go out on Monday, assuming Abigail gives it the thumbs up, I'm
happy to go with this solution for now and check with @tadast when he's back
from holiday.