Fix leave date in Paternity calculator
DRYing out logic in the commit titled "Respect user provided input for
paternity leave duration" has raised my attention to the leave_end_date
calculation that required to add an extra day to the value returned by the
calculator in order to keep regression tests unchanged for the text generated
using this value. I've checked with HMRC and they have confirmed this
calculation was incorrect.