Extract SmartAnswer::I18nRenderer from QuestionPresenter
We're planning to ditch the i18n translation mechanism, so I want to separate
the code responsible for this as much as possible. The new `I18nRenderer` class
is intended to be roughly equivalent to the `ErbRenderer` class.

Unfortunately at this stage the new renderer has rather a large API (i.e. the
five methods that need delegating), but hopefully I can improve on that in
later commits.