Avoid use of respond_to? in _current_question.html.erb
The `_current_question.html.erb` partial template renders both smart answer
questions and smartdown questions (the latter via an adapter layer). Previously
the smart answer `QuestionPresenter` didn't implement `#post_body`, but by
implementing the method (to return `nil`), we can make the behaviour more
explicit and rely on polymorphism instead of the condition in the partial,
this simplifying the code.